Moved Models to separate folder
